update user account

+++ server/db/seed.js
@@ -17,44 +17,22 @@ files.forEach(file => {
   })
 })
 
-const Twitter = require('twitter');
-const env = require('../env')
-const twitter_users = require('../twitter_users')
- 
-var client = new Twitter({
-  consumer_key: process.env.TWITTER_CONSUMER_KEY || env.TWITTER_CONSUMER_KEY,
-  consumer_secret: process.env.TWITTER_CONSUMER_SECRET || env.TWITTER_CONSUMER_SECRET,
-  access_token_key: process.env.TWITTER_ACCESS_TOKEN || env.TWITTER_ACCESS_TOKEN,
-  access_token_secret: process.env.TWITTER_TOKEN_SECRET || env.TWITTER_TOKEN_SECRET 
-});
- 
-var params = {screen_name: 'drinkingChai'};
-client.get('statuses/user_timeline', params)
-
-const fetchTwitterData = () => {
-  return Promise.all(twitter_users.map(user => User.create({ email: user, password: 'jot' })))
-    .then(users => {
-      return Promise.all(users.map(user => {
-        return client.get('statuses/user_timeline', { screen_name: user.email })
-          .then(tweets => {
-            return Promise.all(tweets.map(tweet => {
-              const { text, created_at } = tweet
-              return new Promise(function(resolve) {
-                setTimeout(() => {
-                  resolve(Thought.storeAndCluster({ text, created_at, userId: user.id }))
-                }, 3500)
-              })
-            }))
-          })
-      }))
+// seed prompt
+const prompt = 'Seed the database? (Y/N) '
+process.stdin.on('data', function (data) {
+  const cmd = data.toString().trim()
+  if (cmd == 'Y') {
+    conn.sync({ force: true })
+      .then(() => Promise.all(documents.map(doc => MachineData.create(doc))))
+      .then(() => User.create({ email: 'teatocode@gmail.com', password: 'jot' }))
+      .then(() => {
+        console.log('db seeded')
+        conn.close()
+        process.exit(0)
     })
-} 
+  } else {
+    process.exit(0)
+  }
+})
 
-conn.sync({ force: true })
-  .then(() => Promise.all(documents.map(doc => MachineData.create(doc))))
-  .then(() => User.create({ email: 'teatocode@gmail.com', password: 'jot' }))
-  .then(() => fetchTwitterData())
-  .then(() => {
-    console.log('db seeded')
-    conn.close()
-  })
+process.stdout.write(prompt)
